moved text coloring logic to utils
